tests/wpt/css-tests/css-regions-1_dev/xhtml1/regions-mouse-events-004.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title>CSS Regions: onmouseup events inside region</title>
        <link href="mailto:dalcala@adobe.com" rel="author" title="David Alcala" />
        <link href="http://www.w3.org/TR/css3-regions/#the-flow-into-property" rel="help" />
        <link href="http://www.w3.org/TR/css3-regions/#flow-from" rel="help" />
        <link href="http://www.w3.org/TR/css3-regions/#relation-to-document-events" rel="help" />
        <meta content="CSS regions module does not alter the normal processing of events in the document
            tree. Therefore, onmouseup events inside a named flow should function correctly." name="assert" />
        <meta content="dom interact" name="flags" />
        <link href="support/css/regions-mouse-events.css" type="text/css" rel="stylesheet" />
        <link href="/resources/testharness.css" type="text/css" rel="stylesheet" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="support/js/regions-mouse-events-test-helper.js"></script>
    </head>
    <body>
        <p id="msg">Mouse up on the black square (Press the mouse button down somewhere outside the black square.  Then drag the mouse to the black square.  Release the mouse button while on the black square.)</p>
        <div id="content">
            <div class="block"></div>
            <div class="block"></div>
            <div class="block" id="target-block"></div>
        </div>
        <div class="region"></div>
        <div class="region"></div>
        <div class="region"></div>
        <br />
        <div id="log"></div>
    
     <script type="text/javascript">

        if(getLeftPosition("target-block") !== 0) {
            var test1 = async_test("onmouseup event fired inside region", {timeout: testTimeout});
            test1.step(function () {
                var testTarget = document.getElementById("target-block");
                testTarget.onmouseup = test1.step_func(function (evt) {

                    /* Verify the target inside the region has the correct ID */
                    test(function() {
                        assert_equals(evt.target.id, "target-block");
                    }, "Target that mouseup was performed on inside region has correct ID");

                    test1.done();
                });
                mouseUp("target-block");
            });
            completionCallback();
        }
        else {
            test(function () {
                assert_true(false);
            }, "Regions are not enabled");
        }

    </script>
</body></html>
